Relocate Inches class to new namespace and file
The Inches class was moved from the `Calcify.Math.Conversion.Length`
namespace to the `Calcify.Classes.Math.Conversion.Length` namespace.
This involved removing the class from `Math.cs` and adding it to a
new file, `Inches.cs`. The `Calcify.csproj` file was updated to
include the new file. No changes were made to the logic or methods
of the Inches class.